Refactoring and cleaning up imports to support Windows install
