+++ openfl/federated/task/runner_gandlf.py
@@ -11,7 +11,7 @@ import torch as pt
 from typing import Union
 import yaml
 
-from openfl.utilities import split_tensor_dict_for_holdouts
+from openfl.utilities.split import split_tensor_dict_for_holdouts
 from openfl.utilities import TensorKey
 
 from .runner import TaskRunner
