+++ openfl/federated/task/runner_keras.py
@@ -13,7 +13,7 @@ import numpy as np
 
 from openfl.utilities import change_tags
 from openfl.utilities import Metric
-from openfl.utilities import split_tensor_dict_for_holdouts
+from openfl.utilities.split import split_tensor_dict_for_holdouts
 from openfl.utilities import TensorKey
 from .runner import TaskRunner
