+++ openfl/federated/task/task_runner.py
@@ -7,7 +7,7 @@ from logging import getLogger
 import numpy as np
 
 from openfl.utilities import change_tags
-from openfl.utilities import split_tensor_dict_for_holdouts
+from openfl.utilities.split import split_tensor_dict_for_holdouts
 from openfl.utilities import TensorKey
