+++ openfl/interface/cli.py
@@ -14,9 +14,34 @@ from click import option
 from click import pass_context
 from click import style
 import time
+import logging
 
-from openfl.utilities import add_log_level
 import sys
+
+def add_log_level(level_name, level_num, method_name=None):
+    """
+    Add a new logging level to the logging module.
+
+    Args:
+        level_name: name of log level.
+        level_num: log level value.
+        method_name: log method wich will use new log level (default = level_name.lower())
+
+    """
+    if not method_name:
+        method_name = level_name.lower()
+
+    def log_for_level(self, message, *args, **kwargs):
+        if self.isEnabledFor(level_num):
+            self._log(level_num, message, args, **kwargs)
+
+    def log_to_root(message, *args, **kwargs):
+        logging.log(level_num, message, *args, **kwargs)
+
+    logging.addLevelName(level_num, level_name)
+    setattr(logging, level_name, level_num)
+    setattr(logging.getLoggerClass(), method_name, log_for_level)
+    setattr(logging, method_name, log_to_root)
 
 
 def setup_logging(level='info', log_file=None):
