+++ openfl/interface/plan.py
@@ -52,7 +52,7 @@ def initialize(context, plan_config, cols_config, data_config,
 
     from openfl.federated import Plan
     from openfl.protocols import utils
-    from openfl.utilities import split_tensor_dict_for_holdouts
+    from openfl.utilities.split import split_tensor_dict_for_holdouts
     from openfl.utilities.utils import getfqdn_env
 
     for p in [plan_config, cols_config, data_config]:
