+++ setup.py
@@ -94,7 +94,6 @@ setup(
         'openfl.interface.aggregation_functions.experimental',
         'openfl.component.aggregator',
         'openfl.component.assigner',
-        'openfl.component.ca',
         'openfl.component.collaborator',
         'openfl.component.director',
         'openfl.component.envoy',
@@ -123,6 +122,7 @@ setup(
         'openfl.transport',
         'openfl.transport.grpc',
         'openfl.utilities',
+        'openfl.utilities.ca',
         'openfl.utilities.data_splitters',
         'openfl.utilities.fedcurv',
         'openfl.utilities.fedcurv.torch',
